Merge branch 'charts' into main
Charts were moved to their own git repo.
But I think it makes sense to keep them in the same
main deployment one.

They are still reusable by others here.
But this allows docs to be more central, and also
changes to be made more atomically.

This merge commit also changes the directory that the
charts were in (was base, now in charts)

This also updates the usage of the charts in the repo