mediawiki: introduce web pod autoscaler
This commit introduces a pod autoscaler for the mediawiki web
pod. We have progressively increased this from 1 to 3 replicas
over time so being able to have this dynamically scale would give us
some headroom to hopefully avoid future issues around cpu
exhaustion.

Enabling this is made configurable using a similar pattern of
values to ingress-nginx[1]. Leaving it configurable keeps the
possibility to not scale locally as well as turn it off if we find the
autoscaler to be problematic.

80% cpu was chosen as it is the default for horizontal pod
autoscaling.

[1] https://github.com/kubernetes/ingress-nginx/blob/main/charts/ingress-nginx/values.yaml#L393