linting: update ct version to v3.4.0
This is the version used by the github action version
currently in use (v2.1.0)

This commit also removes what appears now to be a
deprecated `timeout` flag passed to helm which was causing
the updated version to fail to run.

This commit also ensures that the docker user running ct matches
the local user rather than running as root and causing
git issues.