Increase chart versions to pass linting
This commit also introduces using a variable in the
app-of-apps chart to specify versions of the individual
helm charts we wish to apply.

This should prevent requiring a new helmfile apply
needing to be run to bump versions of charts.

It should also prevent us needed to release a new
argocd-config chart when we want to bump a released chart version