argocd: add redis (#171)
* add redis Application manifest and bitnami source

* use temp deploy branch

* update chart name

* redis git test

* fix yaml

* update repo url

* use HEAD

* update

* add bitnami chart repo

* bump chart version

* bump app-of-apps chart version

* bump again

* remove new api version

* bump chart version

* Increase chart versions to pass linting

This commit also introduces using a variable in the
app-of-apps chart to specify versions of the individual
helm charts we wish to apply.

This should prevent requiring a new helmfile apply
needing to be run to bump versions of charts.

It should also prevent us needed to release a new
argocd-config chart when we want to bump a released chart version

* inject chart versions into app-of-apps

* ensure using HEAD deploy

* tweak indenting

* tweak indenting back

---------

Co-authored-by: Thomas Arrow <tarrow@users.noreply.github.com>
Co-authored-by: Thomas Arrow <thomas.arrow@wikimedia.de>